Conflicts:
	doc/help.txt
	package.json
	spec/install.spec.js
	spec/platforms/ios.spec.js
	spec/prepare.spec.js
	src/fetch.js
	src/install.js
	src/platforms/ios.js
	src/prepare.js
	src/util/config-changes.js

+++ src/platforms/android.js
@@ -20,7 +20,6 @@
 var fs = require('fs')  // use existsSync in 0.6.x
    , path = require('path')
    , common = require('./common')
-   , config_changes = require('../util/config-changes')
    , xml_helpers = require(path.join(__dirname, '..', 'util', 'xml-helpers'));
 
 module.exports = {
@@ -48,6 +47,8 @@ module.exports = {
     "source-file":{
         install:function(source_el, plugin_dir, project_dir, plugin_id) {
             var dest = path.join(source_el.attrib['target-dir'], path.basename(source_el.attrib['src']));
+            var target_path = common.resolveTargetPath(project_dir, dest);
+            if (fs.existsSync(target_path)) throw new Error('"' + target_path + '" already exists!');
             common.copyFile(plugin_dir, source_el.attrib['src'], project_dir, dest);
         },
         uninstall:function(source_el, project_dir, plugin_id) {
@@ -55,28 +56,44 @@ module.exports = {
             common.deleteJava(project_dir, dest);
         }
     },
+    "header-file": {
+        install:function(source_el, plugin_dir, project_dir, plugin_id) {
+            require('../../plugman').emit('verbose', 'header-file.install is not supported for android');
+        },
+        uninstall:function(source_el, project_dir, plugin_id) {
+            require('../../plugman').emit('verbose', 'header-file.uninstall is not supported for android');
+        }
+    },
     "lib-file":{
-        install:function(lib_el, plugin_dir, project_dir) {
+        install:function(lib_el, plugin_dir, project_dir, plugin_id) {
             var src = lib_el.attrib.src;
             var dest = path.join("libs", path.basename(src));
             common.copyFile(plugin_dir, src, project_dir, dest);
         },
-        uninstall:function(lib_el, project_dir) {
+        uninstall:function(lib_el, project_dir, plugin_id) {
             var src = lib_el.attrib.src;
             var dest = path.join("libs", path.basename(src));
             common.removeFile(project_dir, dest);
         }
     },
     "resource-file":{
-        install:function(el, plugin_dir, project_dir) {
+        install:function(el, plugin_dir, project_dir, plugin_id) {
             var src = el.attrib.src;
             var target = el.attrib.target;
             require('../../plugman').emit('verbose', 'Copying resource file ' + src + ' to ' + target);
-            common.copyFile(plugin_dir, src, project_dir, target);
+            common.copyFile(plugin_dir, src, project_dir, path.normalize(target));
         },
-        uninstall:function(el, project_dir) {
+        uninstall:function(el, project_dir, plugin_id) {
             var target = el.attrib.target;
-            common.removeFile(project_dir, target);
+            common.removeFile(project_dir, path.normalize(target));
+        }
+    },
+    "framework": {
+        install:function(source_el, plugin_dir, project_dir, plugin_id) {
+            require('../../plugman').emit('verbose', 'framework.install is not supported for android');
+        },
+        uninstall:function(source_el, project_dir, plugin_id) {
+            require('../../plugman').emit('verbose', 'framework.uninstall is not supported for android');
         }
     },
     "proguard-config":{
@@ -99,7 +116,7 @@ module.exports = {
     },
     "android-application":{
         install:function(application_el, plugins_dir, project_dir) {
-            var platformConfig = config_changes.get_platform_json(plugins_dir, 'android'),
+            var platformConfig = getConfigChanges().get_platform_json(plugins_dir, 'android'),
                 extraConfig = platformConfig.extra_config;
 
             var applicationClass = application_el.attrib['name'];
@@ -112,12 +129,12 @@ module.exports = {
             var originApplicationClass = applicationTag.attrib['android:name'] || '';
             extraConfig['android-application'] = {};
             extraConfig['android-application'][applicationClass] = originApplicationClass;
-            config_changes.save_platform_json(platformConfig, plugins_dir, 'android');
+            getConfigChanges().save_platform_json(platformConfig, plugins_dir, 'android');
             applicationTag.attrib['android:name'] = applicationClass;
             fs.writeFileSync(manifestXml, mDoc.write({indent: 4}), 'utf-8');
         },
         uninstall:function(application_el, plugins_dir, project_dir) {
-            var platformConfig = config_changes.get_platform_json(plugins_dir, 'android'),
+            var platformConfig = getConfigChanges().get_platform_json(plugins_dir, 'android'),
                 extraConfig = platformConfig.extra_config;
 
             var applicationClass = application_el.attrib['name'];
@@ -133,12 +150,12 @@ module.exports = {
             fs.writeFileSync(manifestXml, mDoc.write({indent: 4}), 'utf-8');
 
             delete extraConfig['android-application'];
-            config_changes.save_platform_json(platformConfig, plugins_dir, 'android');
+            getConfigChanges().save_platform_json(platformConfig, plugins_dir, 'android');
         }
     },
     "root-activity":{
         install:function(activity_el, plugins_dir, project_dir) {
-            var platformConfig = config_changes.get_platform_json(plugins_dir, 'android'),
+            var platformConfig = getConfigChanges().get_platform_json(plugins_dir, 'android'),
                 extraConfig = platformConfig.extra_config;
             var rootActivityCanonicalName = activity_el.attrib['name'],
                 rootActivityName = getLastPart(rootActivityCanonicalName);
@@ -158,10 +175,10 @@ module.exports = {
             content = content.replace(originRootActivityCanonicalName, rootActivityCanonicalName)
                 .replace(originRootActivityName, rootActivityName);
             fs.writeFileSync(projectActivityPath, content, 'utf-8');
-            config_changes.save_platform_json(platformConfig, plugins_dir, 'android');
+            getConfigChanges().save_platform_json(platformConfig, plugins_dir, 'android');
         },
         uninstall:function(activity_el, plugins_dir, project_dir) {
-            var platformConfig = config_changes.get_platform_json(plugins_dir, 'android'),
+            var platformConfig = getConfigChanges().get_platform_json(plugins_dir, 'android'),
                 extraConfig = platformConfig.extra_config;
             var rootActivityCanonicalName = activity_el.attrib['name'],
                 rootActivityName = getLastPart(rootActivityCanonicalName);
@@ -176,10 +193,14 @@ module.exports = {
             content = content.replace(rootActivityCanonicalName, originRootActivityCanonicalName)
                 .replace(rootActivityName, originRootActivityName);
             fs.writeFileSync(projectActivityPath, content, 'utf-8');
-            config_changes.save_platform_json(platformConfig, plugins_dir, 'android');
+            getConfigChanges().save_platform_json(platformConfig, plugins_dir, 'android');
         }
     }
 };
+
+function getConfigChanges() {
+    return require('../util/config-changes');
+}
 
 function findRootActivitySimpleName(fileData, simpleMainActivityName) {
     return fileData.match(new RegExp('public\\s+class\\s+' + simpleMainActivityName + '\\s+extends\\s+(\\w+)'))[1]; // i.e. XFaceMainActivity
